updated delete message view, can only delete message if message.user_id matches logged in user.id

+++ app.py
@@ -4,7 +4,7 @@ import requests
 import geocoder
 
 from secrets_1 import SECRET, API_KEY
-from flask import Flask, render_template, request, flash, redirect, session, g
+from flask import Flask, render_template, request, flash, redirect, session, g, abort
 from flask_debugtoolbar import DebugToolbarExtension
 from sqlalchemy.exc import IntegrityError
 
@@ -252,6 +252,40 @@ def edit_profile():
     return render_template('users/edit.html', form=form, user_id=user.id)
 
 
+@app.route('/users/<int:user_id>/likes', methods=["GET"])
+def show_likes(user_id):
+    """route to display all liked messages by curr user"""
+    if not g.user:
+        flash("Access password restricted", "danger")
+        return redirect("/")
+    
+    user = User.query.get_or_404(user_id)
+    return render_template("users/likes.html", user=user, likes=user.likes)
+
+@app.route('/messages/<int:message_id>/like', methods=["POST"])
+def add_like(message_id):
+    """Toggle a liked message for the currently-logged in user"""
+    if not g.user:
+        flash("Access password Restricted", "danger")
+        return redirect('/')
+    
+    liked_message = Message.query.get_or_404(message_id)
+
+    if liked_message.user_id == g.user.id:
+        return abort(403)
+    
+    user_likes = g.user.likes
+
+    if liked_message in user_likes:
+        g.user.likes = [like for like in user_likes if like != liked_message]
+    else:
+        g.user.likes.append(liked_message)
+
+    db.session.commit()
+
+    return redirect('/')
+
+
 @app.route('/users/delete', methods=["POST"])
 def delete_user():
     """Delete user."""
@@ -266,7 +300,6 @@ def delete_user():
     db.session.commit()
 
     return redirect("/signup")
-
 
 ##############################################################################
 # Messages routes:
@@ -309,8 +342,12 @@ def messages_destroy(message_id):
     if not g.user:
         flash("Access unauthorized.", "danger")
         return redirect("/")
-
+    
     msg = Message.query.get(message_id)
+    if msg.user_id != g.user.id:
+        flash("Access restricted", "danger")
+        return redirect("/")
+    
     db.session.delete(msg)
     db.session.commit()
